# watchdog/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "description": "Service HealthCheck",
                "tags": [
                    "HealthCheck"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/clients": {
            "get": {
                "description": "Get Clients Status",
                "tags": [
                    "Get Hosts"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/config": {
            "get": {
                "description": "Get watchdog configuration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get Config"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.YamlConfig"
                        }
                    }
                }
            },
            "post": {
                "description": "Update watchdog configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update Config"
                ],
                "parameters": [
                    {
                        "description": "YAML Configuration",
                        "name": "model.yamlConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.YamlConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.YamlConfig"
                        }
                    }
                }
            }
        },
        "/hosts": {
            "get": {
                "description": "List host",
                "tags": [
                    "Get Hosts"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/list": {
            "get": {
                "description": "List miners in each host",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List Miners by host"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Host IP",
                        "name": "host",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.MinerInfoVO"
                            }
                        }
                    }
                }
            }
        },
        "/toggle": {
            "get": {
                "description": "Get Alert Status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get Alert Status"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            },
            "post": {
                "description": "Set Alert Status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Set Alert Status"
                ],
                "parameters": [
                    {
                        "description": "Alert Toggle Status",
                        "name": "model.AlertToggle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AlertToggle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AlertToggle"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core.MinerInfo": {
            "type": "object",
            "properties": {
                "cinfo": {
                    "$ref": "#/definitions/model.Container"
                },
                "conf": {
                    "$ref": "#/definitions/model.MinerConfigFile"
                },
                "minerStat": {
                    "$ref": "#/definitions/model.MinerStat"
                },
                "name": {
                    "type": "string"
                },
                "signatureAcc": {
                    "type": "string"
                }
            }
        },
        "model.AlertToggle": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        },
        "model.Container": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "cpu_percent": {
                    "type": "number"
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "image_id": {
                    "type": "string"
                },
                "mem_usage": {
                    "type": "integer"
                },
                "memory_percent": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.HostItem": {
            "type": "object",
            "properties": {
                "capath": {
                    "description": "/etc/docker/127.0.0.1/ca.pem",
                    "type": "string"
                },
                "certPath": {
                    "description": "/etc/docker/127.0.0.1/cert.pem",
                    "type": "string"
                },
                "ip": {
                    "description": "host ip",
                    "type": "string"
                },
                "keyPath": {
                    "description": "/etc/docker/127.0.0.1/key.pem",
                    "type": "string"
                },
                "port": {
                    "description": "docker api port",
                    "type": "string"
                }
            }
        },
        "model.MinerConfigFile": {
            "type": "object",
            "properties": {
                "boot": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "earningsAcc": {
                    "type": "string"
                },
                "mnemonic": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "rpc": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stakingAcc": {
                    "type": "string"
                },
                "teeList": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "useCpu": {
                    "type": "integer"
                },
                "useSpace": {
                    "type": "integer"
                },
                "workspace": {
                    "type": "string"
                }
            }
        },
        "model.MinerStat": {
            "type": "object",
            "properties": {
                "collaterals": {
                    "type": "string"
                },
                "debt": {
                    "type": "string"
                },
                "declaration_space": {
                    "description": "unit: TiB",
                    "type": "string"
                },
                "idle_space": {
                    "type": "string"
                },
                "is_punished": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "boolean"
                        }
                    }
                },
                "lock_space": {
                    "description": "upload file allocated to this miner but not get a proof from tee yet, it can be serviceSpace after get proof from tee",
                    "type": "string"
                },
                "peer_id": {
                    "type": "string"
                },
                "reward_issued": {
                    "type": "string"
                },
                "service_space": {
                    "type": "string"
                },
                "status": {
                    "description": "positive, exit, frozen, unready(register on chain but no get a tag from tee)",
                    "type": "string"
                },
                "total_reward": {
                    "type": "string"
                }
            }
        },
        "model.YamlConfig": {
            "type": "object",
            "properties": {
                "alert": {
                    "type": "object",
                    "properties": {
                        "email": {
                            "type": "object",
                            "properties": {
                                "receiver": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "smtp_account": {
                                    "type": "string"
                                },
                                "smtp_endpoint": {
                                    "type": "string"
                                },
                                "smtp_password": {
                                    "type": "string"
                                },
                                "smtp_port": {
                                    "type": "integer"
                                }
                            }
                        },
                        "enable": {
                            "type": "boolean"
                        },
                        "webhook": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "hosts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.HostItem"
                    }
                },
                "scrapeInterval": {
                    "type": "integer"
                }
            }
        },
        "service.MinerInfoVO": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "minerInfoList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.MinerInfo"
                    }
                }
            }
        }
    }
}